Split up test utils
