+++ anndata/testing/__init__.py
@@ -1,0 +1,2 @@
+from .equality import assert_equal, assert_adata_equal
+from .exceptions import report_name
