+++ anndata/tests/test_backed_sparse.py
@@ -5,7 +5,8 @@ from scipy import sparse
 
 import anndata as ad
 from anndata._core.sparse_dataset import SparseDataset
-from anndata.tests.helpers import assert_equal, subset_func
+from anndata.testing import assert_equal
+from anndata.tests.helpers import subset_func
 
 subset_func2 = subset_func
