+++ anndata/tests/test_base.py
@@ -8,7 +8,8 @@ from scipy import sparse as sp
 from scipy.sparse import csr_matrix, issparse
 
 from anndata import AnnData
-from anndata.tests.helpers import assert_equal, gen_adata
+from anndata.testing import assert_equal
+from anndata.tests.helpers import gen_adata
 
 
 # some test objects that we use below
